rework latest_hockeyapp_notes action for #806
